refactor(#20): Modernize compiler flags
- Add filtered -Wextra, -Wnarrowing, -Wshadow, -Wfloat-conversion

- Fix bug with prototype entity initialization with radios

- Switch all OpenGL calls using the "f" variant to the "d" variant, as when
  ARGoS is built for the simulator, Real=double, following the principle of
  least surprise.

- Execute for -Wuseless-cast. Not added as part of the build flags, because it
  throws tons of warnings in TinyXML and Bullet.

+++ src/core/simulator/simulator.cpp
@@ -379,7 +379,7 @@ namespace argos {
                                          "length",
                                          fExpLength,
                                          0.0f);
-         m_unMaxSimulationClock = fExpLength * unTicksPerSec;
+         m_unMaxSimulationClock = static_cast<UInt32>(fExpLength * unTicksPerSec);
          LOG << "[INFO] Total experiment length in clock ticks = "
              << (m_unMaxSimulationClock ? ToString(m_unMaxSimulationClock) : "unlimited")
              << std::endl;
