+++ src/plugins/simulator/media/rab_medium.cpp
@@ -42,9 +42,9 @@ namespace argos {
          if(strPosIndexMethod == "grid") {
             size_t punGridSize[3];
             if(!NodeAttributeExists(t_tree, "grid_size")) {
-               punGridSize[0] = cArenaSize.GetX();
-               punGridSize[1] = cArenaSize.GetY();
-               punGridSize[2] = cArenaSize.GetZ();
+               punGridSize[0] = static_cast<size_t>(cArenaSize.GetX());
+               punGridSize[1] = static_cast<size_t>(cArenaSize.GetY());
+               punGridSize[2] = static_cast<size_t>(cArenaSize.GetZ());
             }
             else {
                std::string strPosGridSize;
